add registration form, login, logout for users

+++ beautypal/views.py
@@ -1,14 +1,66 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
 
-#rooms = [
+# rooms = [
  #   {'id': 1, 'name': 'Lets talk skincare!'},
   #  {'id': 2, 'name': 'Best haircare chat'},
    # {'id': 3, 'name': 'Makeup and color'},
-#]
+# ]
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+
+        else:
+            messages.error(request, 'Username or password does not exist')
+
+    context = {'page': page}
+    return render(request, 'beautypal/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    return render(request, 'beautypal/login_register.html', {'form' : form})
 
 
 def home(request):
@@ -29,6 +81,8 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'beautypal/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -46,9 +100,13 @@ def navbar(request):
     return render(request, 'navbar.html')
     
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You need permission for this task!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -60,8 +118,12 @@ def updateRoom(request, pk):
     return render(request, 'beautypal/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You need permission for this task!')
 
     if request.method == 'POST':
         room.delete()
